+++ projects/komponentkartan/src/lib/controls/combobox/combobox.component.ts
@@ -76,7 +76,9 @@ export class ComboboxComponent implements OnChanges, AfterContentInit, AfterView
 
   @HostListener('document:click', ['$event'])
   onDocumentClick(event: any) {
+
     const target = event.target || event.srcElement || event.currentTarget;
+    
     if ((this.elementRef.nativeElement && !this.elementRef.nativeElement.contains(target)) && this.expanded) {
       this.onChange(null);
       this.collapse(false);
@@ -269,6 +271,8 @@ export class ComboboxComponent implements OnChanges, AfterContentInit, AfterView
       return;
     } else if (event.relatedTarget !== null  && (event.relatedTarget as HTMLElement).id === 'page-content-focus') { // otherwise the pagefocus will steal event and collapse
       return;
+    } else if (event.relatedTarget !== null  && (event.relatedTarget as HTMLElement).tagName === 'VGR-EDITABLE-TABLE-COLUMN') { // otherwise the the combobox in an editable-table will alway collapse on click
+      return;
     }
 
     this.hasFocus = false;
@@ -280,7 +284,6 @@ export class ComboboxComponent implements OnChanges, AfterContentInit, AfterView
       } else {
         this.setSelectedState(this.value);
       }
-
       this.collapse(false);
     } else {
       this.deselectItems();