+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Change-log](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+
+# 9.0.7 (2022-03-08) Stable
+- Added new editable-table component
+- Returned collapse to compbobox and added new terms for blurevent
+- Fixed bug regarding shift+tab move in a grid
+- Fixed that you can tab into a editable-table, next tab gets you to last tabable item in table and then out from table
+- Fixed so a readonly object dont get the red error frame when touched
+- Fixed so you can grab the scroll in a combobox when it lies in another combonent
+- Added check on editmode in the initialisation of the table
+- Moved tabindex for editable-datepicker (but not when disabled)
+
 # 9.0.6 (2022-03-07) Stable
 - Screenreader reads notifications (#557)
 - Added margin-top and bottom when grid is expanded (#549)